# api/openapi.json
{
  "components": {
    "schemas": {
...
      "EchoRequest": {
        "properties": {
          "message": {
            "description": "Message to echo back.",
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "EchoRequest",
        "type": "object"
      },
      "EchoResponse": {
        "properties": {
          "message": {
            "description": "Original message.",
            "title": "Message",
            "type": "string"
          },
          "reversed": {
            "description": "Reversed message.",
            "title": "Reversed",
            "type": "string"
          }
        },
        "required": [
          "message",
          "reversed"
        ],
        "title": "EchoResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "DeviceJWT": {
        "description": "Device-signed ES256 JWT minted by the client. The JWT header must include kid set to the device id, and the aud claim must be h4ckath0n:http.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Hackathon app powered by h4ckath0n",
    "title": "h4ckath0n",
    "version": "0.1.9"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/demo/echo": {
      "post": {
        "description": "Echo back the message along with its reverse.",
        "operationId": "demo_echo_demo_echo_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EchoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EchoResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Demo echo",
        "tags": [
          "demo"
        ]
      }
    },
...
    "/demo/sse": {
      "get": {
        "description": "Authenticated SSE stream that simulates chunked output. Requires a device JWT with aud set to h4ckath0n:sse.",
        "operationId": "demo_sse_demo_sse_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              },
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Event stream with chunk and done events."
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Missing token"
                }
              }
            },
            "description": "Missing or invalid device JWT."
          }
        },
        "summary": "Demo SSE stream",
        "tags": [
          "demo"
        ]
      }
    },
...
  }
}